requirements.txt
winsdk
pypresence
requests
Pillow
pystray
pywebview
